Removed breaking code indentation
I'm no .md file writer, but it seems like > in front of code blocks do mess things up. I've only removed the indentations, but that of course sets the code block a bit out of flow, as it no longer is indented, like the surrounded text. Maybe there is another feature of the editor, that I am not aware of, that can both do code blocks, and indentation at the same lines.